Move New plan tests to NewPlanViewTest
